+++ src/jaclks-internal/javac-base/java/lang/character.cc
@@ -1,4 +1,4 @@
-#include "jaclks/javac-base/lang/character.h"
+#include "jaclks/javac-base/java/lang/character.h"
 
 #include <cctype>