+++ src/jaclks-internal/javac-base/java/lang/system.cc
@@ -1,0 +1,16 @@
+#include "jaclks/javac-base/java/lang/system.h"
+
+#include <chrono>
+
+namespace jaclks::javac_base {
+
+std::int64_t System::CurrentTimeMillis() {
+  using std::chrono::duration_cast;
+  using std::chrono::milliseconds;
+  using std::chrono::system_clock;
+
+  auto now = system_clock::now();
+  return duration_cast<milliseconds>(now.time_since_epoch()).count();
+}
+
+}  // namespace jaclks::javac_base