+++ src/jaclks-internal/javac-base/java/security/md5_digest.cc
@@ -1,0 +1,63 @@
+#include "jaclks-internal/javac-base/java/security/md5_digest.h"
+
+#ifdef USE_EVP_MD5
+#include <openssl/md5.h>
+#endif
+
+namespace jaclks::javac_base {
+
+MD5Digest::MD5Digest() : ctx_{} {
+  md5_init();
+}
+
+MD5Digest::~MD5Digest() {
+#ifdef USE_EVP_MD5
+  EVP_MD_CTX_free(ctx_);
+  ctx_ = nullptr;
+#endif
+}
+
+void MD5Digest::EngineUpdate(const std::uint8_t *data, std::size_t num) {
+  md5_update(data, num);
+}
+
+String MD5Digest::EngineDigest() {
+  return md5_final();
+}
+
+void MD5Digest::EngineReset() {
+  this->~MD5Digest();
+  new (this) MD5Digest();
+}
+
+void MD5Digest::md5_init() {
+#ifdef USE_EVP_MD5
+  ctx_ = EVP_MD_CTX_new();
+  EVP_DigestInit_ex(ctx_, EVP_md5(), nullptr);
+#else
+  MD5_Init(&ctx_);
+#endif
+}
+
+void MD5Digest::md5_update(const std::uint8_t *data, std::size_t len) {
+#ifdef USE_EVP_MD5
+  EVP_DigestUpdate(ctx_, data, len);
+#else
+  MD5_Update(&ctx_, data, len);
+#endif
+}
+
+String MD5Digest::md5_final() {
+  constexpr std::size_t kDigestSize = MD5_DIGEST_LENGTH;
+  unsigned char digest[MD5_DIGEST_LENGTH];
+
+#ifdef USE_EVP_MD5
+  unsigned int len;
+  EVP_DigestFinal_ex(ctx_, digest, &len);
+#else
+  MD5_Final(digest, &ctx_);
+#endif
+  return String{reinterpret_cast<const char *>(digest), kDigestSize};
+}
+
+}  // namespace jaclks::javac_base