+++ CHANGELOG.md
@@ -19,6 +19,9 @@ Please mark all change in change log and use the issue from GitHub
 - [#17](https://github.com/BossZou/jaclks/pull/17) Add StartsWith/EndsWith method && write README
 - [#24](https://github.com/BossZou/jaclks/pull/24) Add Trim/Strip/IndexOf/StartsWith/EndsWith methods
 
+### MD5 digest implementation
+- [#44](https://github.com/BossZou/jaclks/pull/44) Add MD5 algorithm for message digest
+
 ## Improvement
 - [#29](https://github.com/BossZou/jaclks/pull/29) Support on windows 2025 x86 platform
 - [#35](https://github.com/BossZou/jaclks/pull/35) Support on windows 11 arm64 and Macos x86/arm platform